uint16_t -> uint8_t

+++ Source/SideScrollerCPP/OpticalFlow.cpp
@@ -13,8 +13,8 @@
 OpticalFlow::OpticalFlow(AHUD* hud, int leftx, int topy, int rows, int cols) : 
 	_leftx(leftx), _topy(topy), _hud(hud),  _rows(rows), _cols(cols)
 {
-    _imgcurr = new uint16_t [rows*cols];
-    _imgprev = new uint16_t [rows*cols];
+    _imgcurr = new uint8_t [rows*cols];
+    _imgprev = new uint8_t [rows*cols];
 
 	_ready = false;
 }
@@ -29,7 +29,7 @@ void OpticalFlow::perform(uint8_t* imagergb)
 {
     // RGB->gray formula from https ://www.johndcook.com/blog/2009/08/24/algorithms-convert-color-grayscale/
 	for (int k = 0; k < _rows*_cols; ++k) {
-		_imgcurr[k] = (uint16_t)(0.21 * imagergb[3 * k] + 0.72 * imagergb[3 * k + 1] + 0.07 * imagergb[3 * k + 2]);
+		_imgcurr[k] = (uint8_t)(0.21 * imagergb[3 * k] + 0.72 * imagergb[3 * k + 1] + 0.07 * imagergb[3 * k + 2]);
 	}
 
 	// Fake up optical flow with a line for now
